Extract orion state to its own module
